requirements.txt
torch
numpy
google-generativeai
transformers
scikit-learn
python-dotenv

flask
flask-cors
torch
transformers
numpy
scikit-learn
google-generativeai
python-dotenv
